test: move the logs --tail test into cli3 to stay under the line limit
The new test pushed cli2.rs past 500 lines; split it into a cli3 module.
