+++ tests/engine_integration.rs
@@ -56,6 +56,8 @@ mod watch_tests;
 mod cli1;
 #[path = "engine_integration/cli2.rs"]
 mod cli2;
+#[path = "engine_integration/cli3.rs"]
+mod cli3;
 #[path = "engine_integration/create_ls.rs"]
 mod create_ls;
 #[path = "engine_integration/scale.rs"]
